+++ CHANGELOG.md
@@ -9,6 +9,7 @@ a concise history of changes made.
 - Added the 'enabled' field to Observer so loading Checklists can be blocked.
 - Added caching to the Latest, Weekly and Monthly News pages.
 - Added date range fields to the Checklists search panel.
+- Added filter for hotspots to the Checklists search panel.
 
 ## 2025.05.21
 - Added site logo.